@@ -1,20 +1,20 @@
 {
   "name": "weather-dashboard",
-  "version": "0.1.0",
+  "version": "1.0.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^6.1.4",
-    "@testing-library/react": "^14.1.0",
-    "@testing-library/user-event": "^14.5.1",
-    "bootstrap": "^5.3.2",
+    "@testing-library/jest-dom": "^6.5.0",
+    "@testing-library/react": "^16.0.1",
+    "@testing-library/user-event": "^14.5.2",
+    "bootstrap": "^5.3.3",
     "jquery": "^3.7.1",
-    "justgage": "^1.6.1",
-    "moment": "^2.29.4",
+    "justgage": "^1.7.0",
+    "moment": "^2.30.1",
     "raphael": "^2.3.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-scripts": "5.0.1",
-    "web-vitals": "^3.5.0"
+    "web-vitals": "^4.2.3"
   },
   "scripts": {
     "start": "react-scripts start",
