@@ -1,29 +1,22 @@
-import {Component} from "react";
+import React from "react";
 
-class ForecastWidget extends Component {
-    shouldComponentUpdate(nextProps, nextState, nextContext) {
-        return true;
+export default function ForecastWidget({day, forecast}) {
+    if (forecast == null) {
+        return null;
     }
-
-    render() {
-        if (this.props.forecast == null) {
-            return null;
-        }
-        return (
-            <div className="col-sm-4">
-                <h2 id="day">{this.props.day}</h2>
-                <div className="icon-wrapper" id="icon">
-                    <img src={"./img/" + this.props.forecast.iconD + ".png"} height="135" width="135" alt=""/>
-                    {this.props.forecast.iconN !== null ? (
-                        <img src={"./img/" + this.props.forecast.iconN + ".png"} height="135" width="135" alt=""/>
-                    ) : (<img alt=""/>)}
-                </div>
-                <h3 className="weather">{this.props.forecast.weather}</h3>
-                <h3 className="temp">{this.props.forecast.temp}</h3>
-                <h3 className="wind">{this.props.forecast.wind}</h3>
+    return (
+        <div className="col-sm-4">
+            <h2 id="day">{day}</h2>
+            <div className="icon-wrapper" id="icon">
+                <img src={"./img/" + forecast.iconDay + ".png"} height="135" width="135" alt=""/>
+                {forecast.iconNight !== null ? (
+                    <img src={"./img/" + forecast.iconNight + ".png"} height="135" width="135" alt=""/>
+                ) : (<img alt=""/>)}
             </div>
-        );
-    }
-}
-
-export default ForecastWidget;
+            <h3 className="weather">{forecast.weather}</h3>
+            <h3 className="temp">{forecast.temp}</h3>
+            <h3 className="wind">白天：{forecast.windDay}</h3>
+            <h3 className="wind">晚上：{forecast.windNight}</h3>
+        </div>
+    );
+};
